seperate story definitions
