+++ .storybook/stories/TableFiltering/TableFilteringMedium.stories.tsx
@@ -1,0 +1,53 @@
+import { TableFilteringMedium } from './TableFilteringMedium';
+import { useSortAndFilter } from './lib/useSortAndFilter';
+import { generateTableCaption, useData } from './lib/utils';
+
+export default {
+	title: 'Patterns/Data Filtering (WIP)',
+	parameters: {
+		layout: 'fullscreen',
+	},
+};
+
+export const Medium = () => {
+	const {
+		filters,
+		pagination,
+		resetFilters,
+		removeFilter,
+		setFilters,
+		setPagination,
+		setSort,
+		sort,
+	} = useSortAndFilter();
+
+	const { loading, data, totalPages, totalItems } = useData({
+		filters,
+		pagination,
+		sort,
+	});
+
+	const tableCaption = generateTableCaption({
+		loading,
+		totalItems,
+		pagination,
+	});
+
+	return (
+		<TableFilteringMedium
+			data={data}
+			filters={filters}
+			loading={loading}
+			pagination={pagination}
+			resetFilters={resetFilters}
+			setFilters={setFilters}
+			removeFilter={removeFilter}
+			setPagination={setPagination}
+			setSort={setSort}
+			sort={sort}
+			tableCaption={tableCaption}
+			totalPages={totalPages}
+			totalItems={totalItems}
+		/>
+	);
+};
