+++ .storybook/stories/TableFiltering/TableFilteringSmall.stories.tsx
@@ -1,5 +1,4 @@
 import { TableFilteringSmall } from './TableFilteringSmall';
-import { TableFilteringMedium } from './TableFilteringMedium';
 import { useSortAndFilter } from './lib/useSortAndFilter';
 import { generateTableCaption, useData } from './lib/utils';
 
@@ -38,49 +37,6 @@ export const Small = () => {
 			loading={loading}
 			data={data}
 			tableCaption={tableCaption}
-			totalItems={totalItems}
-		/>
-	);
-};
-
-export const Medium = () => {
-	const {
-		filters,
-		pagination,
-		resetFilters,
-		removeFilter,
-		setFilters,
-		setPagination,
-		setSort,
-		sort,
-	} = useSortAndFilter();
-
-	const { loading, data, totalPages, totalItems } = useData({
-		filters,
-		pagination,
-		sort,
-	});
-
-	const tableCaption = generateTableCaption({
-		loading,
-		totalItems,
-		pagination,
-	});
-
-	return (
-		<TableFilteringMedium
-			data={data}
-			filters={filters}
-			loading={loading}
-			pagination={pagination}
-			resetFilters={resetFilters}
-			setFilters={setFilters}
-			removeFilter={removeFilter}
-			setPagination={setPagination}
-			setSort={setSort}
-			sort={sort}
-			tableCaption={tableCaption}
-			totalPages={totalPages}
 			totalItems={totalItems}
 		/>
 	);
